* Metadata searches move to specific file and out of `__ini__`; trying to get coverage working. * Python 3.7.
