+++ setup.py
@@ -66,7 +66,7 @@ setup(name=NAME,
       license=LICENSE,
       url=URL,
       keywords=KEYWORDS,
-      python_requires='>=3.6',
+      python_requires='>=3.7',
       setup_requires=['pytest-runner'],
       tests_require=modules['required'],
       scripts=[
